SQL
Mar'22

Online retail Orders Analysis
Course SQL

This project is based on the order management functionality of an online retail store in which you are provided with the “orders” database and you are asked some queries related to it. Answers to these queries will help the company in making data-driven decisions that will impact the overall growth of the online retail store.  

Skills and Tools

Joins, Sub Queries, SQL-clauses-statements-conditions, SQLite using DB Browser and MySQL Workbench